LayoutTests/imported/w3c/web-platform-tests/import-maps/data-driven/resolving_data-url-prefix.json-expected.txt

CONSOLE MESSAGE: Error: assert_unreached: Invalid message: FetchError Reached unreachable code


Harness Error (TIMEOUT), message = null

PASS Test helper: fetching and sanity checking test JSON: data-url-prefix.json
TIMEOUT data: URL prefix: should not resolve since you can't resolve relative to a data: URL: foo/bar Test timed out
